feat(deprecated): add deprecated CRUD controllers and services for projects and workspaces to facilitate migration to new structure chore(deprecated): remove unused scaffold sort DTO and update imports to reflect new deprecated structure fix(deprecated): adjust imports in various services and controllers to point to the new deprecated paths for better organization
