+++ server/src/projects/_deprecated/project-id-entity-id.dto.ts
@@ -1,6 +1,9 @@
 import {ApiProperty, OmitType} from '@nestjs/swagger';
 import {IsNumber, Min} from 'class-validator';
 
+/**
+ * @deprecated
+ */
 export class ProjectIdEntityIdDto {
     @IsNumber()
     @Min(1)
@@ -13,6 +16,9 @@ export class ProjectIdEntityIdDto {
     projectId: number;
 }
 
+/**
+ * @deprecated
+ */
 export class ProjectIdDto extends OmitType(ProjectIdEntityIdDto, [
     'id'
 ] as const) {}
