+++ server/src/scaffold/_deprecated/scaffold-param.ts
@@ -1,5 +1,8 @@
 import {PipeTransform, Type} from '@nestjs/common';
 
+/**
+ * @deprecated
+ */
 export interface ScaffoldParam {
     method?: Array<MethodDecorator>;
     param?: Array<Type<PipeTransform> | PipeTransform>;
