+++ server/src/scaffold/_deprecated/scaffold-read-dto.service.ts
@@ -1,5 +1,8 @@
-import {ScaffoldEntity} from '@/scaffold/services/scaffold-entity.service';
+import {ScaffoldEntity} from '@/scaffold/_deprecated/scaffold-entity.service';
 
+/**
+ * @deprecated
+ */
 export interface ScaffoldReadDtoService<
     TEntity extends ScaffoldEntity,
     TGetDto extends ScaffoldEntity
