--- server/src/scaffold/dtos/scaffold-sort.dto.ts
@@ -1,9 +1,0 @@
-import {
-    IsSortEnum,
-    ScaffoldSort
-} from '@/scaffold/decorators/is-sort-enum.decorator';
-
-export class ScaffoldSortDto {
-    @IsSortEnum()
-    sort: ScaffoldSort = ScaffoldSort.ASC;
-}
