+++ server/src/scaffold/utils/to-human.utils.ts
@@ -1,4 +1,4 @@
-import {ScaffoldEntity} from '@/scaffold/services/scaffold-entity.service';
+import {ScaffoldEntity} from '@/scaffold/_deprecated/scaffold-entity.service';
 import {Type} from '@nestjs/common';
 
 export function toHumanUtils<Entity extends ScaffoldEntity>(
